Optimize: Require functions for better performance
